# decoder: Small speedups