--- a/ogorek.go
+++ b/ogorek.go
@@ -116,6 +116,9 @@ type Decoder struct {
 	// a reusable buffer that can be used by the various decoding functions
 	// functions using this should call buf.Reset to clear the old contents
 	buf bytes.Buffer
+
+	// reusable buffer for readLine
+	line  []byte
 }
 
 // NewDecoder constructs a new Decoder which will decode the pickle stream in r.
@@ -265,21 +268,23 @@ loop:
 	return d.pop()
 }
 
+// readLine reads next line from pickle stream
+// returned line is valid only till next call to readLine
 func (d *Decoder) readLine() ([]byte, error) {
 	var (
-		line     []byte
 		data     []byte
 		isPrefix = true
 		err      error
 	)
+	d.line = d.line[:0]
 	for isPrefix {
 		data, isPrefix, err = d.r.ReadLine()
 		if err != nil {
-			return line, err
+			return d.line, err
 		}
-		line = append(line, data...)
+		d.line = append(d.line, data...)
 	}
-	return line, nil
+	return d.line, nil
 }
 
 // Push a marker
@@ -646,11 +651,13 @@ func (d *Decoder) global() error {
 	if err != nil {
 		return err
 	}
+	smodule := string(module)
 	name, err := d.readLine()
 	if err != nil {
 		return err
 	}
-	d.stack = append(d.stack, Class{Module: string(module), Name: string(name)})
+	sname := string(name)
+	d.stack = append(d.stack, Class{Module: smodule, Name: sname})
 	return nil
 }