@kisielk 2017-04-05
Unfortunate that this has to change the public API, but I guess it was a mistake to make it not return *Decoder in the first place