@navytux 2017-04-06
Yes, it was probably an overlook or thinko. Thanks for merging the patches.