@kisielk 2017-04-06
If you look at the original 2012 commit date it was from when I was a lot less experienced with Go :)