@navytux 2017-04-06
Yes, I understand. I was too only looking at go tour then. Anyway after those 5 years it seems to be appropriate time to correct :)